perf(ios): halve audio sampling rate
Since audio is already downsampled, the sampling rate can be decreased
such that the emulator is sampled roughly proportional to the device's
output rate.

As a performance optimization, decreasing the frequency of calls across
the FFI boundary on the hot path has a considerable effect.

As of this commit, the emulator has audio sampled at 64 KiHz.