perf(ios): avoid uniffi bridge for audio samples
Profiling the emulation look indicates that a significant amount of time
is spent lifting audio samples from Rust to Swift via UniFFI. This
indicates poor quality in the generated bridge code.

To avoid relying on that code on the hot path, we can encode audio
samples with a native type, in this instance `u64`, which is much
cheaper to lift.

Ideally in the future, this optimization will no longer be needed.